+++ examples/ble_server.rs
@@ -63,6 +63,8 @@ fn main() {
 
   ble_advertising.start().unwrap();
 
+  server::ble_gatts_show_local();
+
   let mut counter = 0;
   loop {
     esp_idf_hal::delay::FreeRtos::delay_ms(1000);